update auth controller to include redis auth mgt

+++ services/auth_service/v1/src/controllers/auth.controller.ts
@@ -2,7 +2,7 @@ import mongoose from 'mongoose';
 import { Request, Response, NextFunction } from 'express';
 import {
     deleteAuthFromCacheMemory,
-    getAuthCodes, getAuthTokens,
+    getAuthCodes, getAuthFromCacheMemory, getAuthTokens,
     handleExistingUser, handleUnverifiedUser
 } from '../services/auth.service';
 import { sendEmail } from '../services/email.service';
@@ -78,7 +78,6 @@ const userSignup = async (req: Request, res: Response, next: NextFunction) => {
 
     // Get access token
     const populated_user: UserWithStatus = await user.populate<UserWithStatus>('status')
-    console.log('handleUnverifiedUser')
     return await handleUnverifiedUser(populated_user.toObject(), res);
 };
 
@@ -115,23 +114,35 @@ const verifyUserEmail = async (req: AuthenticatedRequest, res: Response, next: N
 
     // Get user
     const user = req.user
-
-    // if (user.status.isVerified) return next(new BadRequestError('User already verified'));
+    if (user.status.isVerified) return next(new BadRequestError('User already verified'));
 
     // // Check if verification code is correct
-    // const auth_code = await AuthCode.findOne({ user: user._id, });
+    const auth_code = await getAuthFromCacheMemory({
+        auth_class: 'code',
+        type: 'verification',
+        email: req.user.email
+    })
 
-    // if (auth_code?.verification_code !== verification_code) {
-    //     return next(new BadRequestError('Invalid verification code'))
-    // }
+    if (!auth_code || auth_code == verification_code.toString()) {
+        return next(new BadRequestError('Invalid verification code'))
+    }
 
     // // Verify user
-    // await Status.findOneAndUpdate({ user: user._id }, { isVerified: true });
+    await Status.findOneAndUpdate({ user: user._id }, { isVerified: true });
 
-    // await auth_code.updateOne({ verification_code: undefined })
+    // Blacklist access token
+    deleteAuthFromCacheMemory({
+        auth_class: 'token',
+        type: 'verification',
+        email: req.user.email
+    })
 
-    // // Blacklist access token
-    // await BlacklistedToken.create({ token: req.headers.authorization.split(' ')[1] })
+    // Delete verification code
+    deleteAuthFromCacheMemory({
+        auth_class: 'code',
+        type: 'verification',
+        email: req.user.email
+    })
 
     res.status(200).send({
         status: 'success',
@@ -203,20 +214,31 @@ const resetPassword = async (req: AuthenticatedRequest, res: Response, next: Nex
     const { password_reset_code, new_password } = req.body;
 
     // Check if password reset code is correct
-    // const auth_code = await AuthCode.findOne({ user: req.user._id, password_reset_code });
+    const auth_code = await getAuthFromCacheMemory({
+        auth_class: 'code',
+        type: 'password_reset',
+        email: req.user.email
+    })
 
-    // if (!auth_code) return next(new BadRequestError('Invalid password reset code'));
+    if (!auth_code || auth_code != password_reset_code.toString()) {
+        return next(new BadRequestError('Invalid password reset code'));
+    }
 
-    // // Update password
-    // const password = await Password.findOne({ user: req.user._id });
+    // Update password
+    const password = await Password.findOne({ user: req.user._id });
 
-    // password
-    //     ? await password.updatePassword(new_password)
-    //     : next(new InternalServerError('An error occurred'));
+    password
+        ? await password.updatePassword(new_password)
+        : next(new InternalServerError('An error occurred'));
 
     // // Blacklist access token
     await deleteAuthFromCacheMemory({
         auth_class: 'code',
+        type: 'password_reset',
+        email: req.user.email,
+    })
+    await deleteAuthFromCacheMemory({
+        auth_class: 'token',
         type: 'password_reset',
         email: req.user.email,
     })
